# server/openapi.yaml
openapi: 3.0.3
info:
  title: UCC-MCA Intelligence API
  description: |
    AI-powered lead generation system for Merchant Cash Advance providers.
    Analyzes Uniform Commercial Code (UCC) filings to identify businesses
    with active financing and predict MCA likelihood.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/portfolio:
    get:
      tags:
        - Portfolio
      summary: List portfolio companies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: health_grade
          in: query
          schema:
            type: string
            enum:
              - A
              - B
              - C
              - D
              - F
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - funded_date
              - health_score
              - company_name
            default: funded_date
        - name: sort_order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: List of portfolio companies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/portfolio/{id}:
    get:
      tags:
        - Portfolio
      summary: Get portfolio company details
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Portfolio company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioCompany'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/portfolio/stats/summary:
    get:
      tags:
        - Portfolio
      summary: Get portfolio statistics
      responses:
        '200':
          description: Portfolio statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioStats'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===========================================
  # Enrichment Endpoints
  # ===========================================
...
  /api/jobs/{jobId}:
    get:
      tags:
        - Jobs
      summary: Get job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Jobs
      summary: Remove a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  jobId:
                    type: string
                  queueName:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/keys:
    post:
      tags:
        - API Keys
      summary: Create an API key
      description: >
        Mint a new org-scoped API key. The plaintext key is returned ONCE in the
        response and cannot be retrieved again. Admin (JWT) only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                  example: acme-production
                role:
                  type: string
                  enum:
                    - user
                    - admin
                  default: user
                expires_at:
                  type: string
                  format: date-time
                  description: Optional ISO-8601 expiry; omit for a non-expiring key
      responses:
        '201':
          description: Key created; plaintext key returned once
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      key:
                        type: string
                        description: Plaintext key, shown only once
                        example: prk_AbC123...
                      keyPrefix:
                        type: string
                      role:
                        type: string
                      expiresAt:
                        type: string
                        nullable: true
                  meta:
                    type: object
        '400':
          description: Validation failed or no org context
        '403':
          description: Caller is not an admin
    get:
      tags:
        - API Keys
      summary: List API keys
      description: List the org's API keys (metadata only; never returns secrets). Admin (JWT) only.
      responses:
        '200':
          description: The org's API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '403':
          description: Caller is not an admin
  /api/keys/{id}:
    delete:
      tags:
        - API Keys
      summary: Revoke an API key
      description: Revoke an API key by id within the caller's org. Admin (JWT) only.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Key revoked
        '403':
          description: Caller is not an admin
        '404':
          description: Key not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Org-scoped API key (prefixed `prk_`) for programmatic access to the
        on-demand scrape endpoints. May also be sent as `Authorization: Bearer
        prk_…`. Mint and revoke keys via the `/api/keys` endpoints (admin only).

    twilioSignature:
      type: apiKey
      in: header
      name: X-Twilio-Signature
      description: Twilio request signature

    sendgridSignature:
      type: apiKey
      in: header
      name: X-Twilio-Email-Event-Webhook-Signature
      description: SendGrid webhook signature

    plaidSignature:
      type: apiKey
      in: header
      name: Plaid-Verification
      description: Plaid webhook signature
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
            - code
          properties:
            message:
              type: string
            code:
              type: string
            statusCode:
              type: integer
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    PortfolioCompany:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
        funded_date:
          type: string
          format: date
        funded_amount:
          type: number
        health_score:
          type: integer
        health_grade:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - F
    PortfolioListResponse:
      type: object
      properties:
        companies:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioCompany'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    PortfolioStats:
      type: object
      properties:
        total_companies:
          type: integer
        total_funded:
          type: number
        average_health_score:
          type: number
        grade_distribution:
          type: object
          additionalProperties:
            type: integer

    # ===========================================
    # Enrichment Schemas
    # ===========================================
...
    JobDetail:
      type: object
      properties:
        jobId:
          type: string
        queueName:
          type: string
        status:
          type: string
          enum:
            - waiting
            - active
            - completed
            - failed
            - delayed
        progress:
          type: number
        data:
          type: object
        uccProvider:
          type: string
          nullable: true
        returnvalue:
          type: object
        failedReason:
          type: string
        processedOn:
          type: integer
        finishedOn:
          type: integer
...